+++ src/entities/InputBase.ts
