+++ src/entities/InputGap.ts
