+++ src/entities/InputNumberOfHexagons.ts
